requirements.txt
Django
ipython
psycopg2-binary
pillow
pytils
redis
python-dotenv

django4-background-tasks